backend/requirements.txt
Django>=5.0.4,<5.1
psycopg2-binary==2.9.9
psycopg2==2.9.9
python-dotenv==0.20.0
requests == 2.32.3
# pandas == 2.0.3

asgiref
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
python-dotenv